+++ examples/01-basic-entity/ent/contract/sqlc/schema.sql
@@ -3,10 +3,10 @@
 
 -- user table
 CREATE TABLE "user"(
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   email TEXT UNIQUE NOT NULL,
   name TEXT NOT NULL,
-  age INTEGER,
+  age BIGINT,
   uuid TEXT NOT NULL,
   is_admin BOOLEAN NOT NULL,
   created_at TIMESTAMPTZ NOT NULL,
